requirements.txt
rasa
aiohttp
ruamel.yaml
supervisor
sanic-ext<22